refactor: fix asset loading issues and improve asset management structure
