+++ ASSETS_FIXED.md
@@ -1,0 +1,67 @@
+# 🎉 Asset Error Fixed! - Final Status
+
+## ✅ Problem Solved!
+
+The error `libraries/berrio.lua:43: Não foi possível abrir o arquivo: assets/valid_answers.csv` has been **completely fixed**!
+
+### What was the issue?
+- The CSV files were moved to `assets/data/` during modularization
+- But `gameLogic.lua` was still using the old paths: `assets/valid_answers.csv`
+- The Berrio library couldn't find the files at the old location
+
+### ✅ Solutions Applied:
+
+1. **Updated Asset Paths**:
+   - Fixed `gameLogic.lua` to use correct paths: `assets/data/valid_answers.csv`
+   - Added proper asset path management in `assetManager.lua`
+
+2. **Improved Asset Management**:
+   - Added `getGameDataPaths()` function to centralize data file paths
+   - Now all CSV paths are managed through the asset manager
+
+3. **Better Library Organization**:
+   - Moved `berrio.lua` from `libraries/` to `src/libs/`
+   - Updated require path in `main.lua`
+   - Cleaned up empty folders
+
+### 🏗️ Final Clean Project Structure:
+
+```
+berr.io/
+├── main.lua                    # Main game loop
+├── src/
+│   ├── libs/                  # External libraries
+│   │   └── berrio.lua         # Word validation (now in proper location)
+│   ├── modules/               # Game modules
+│   │   ├── config.lua         # Configuration
+│   │   ├── utils.lua          # Utilities
+│   │   ├── ui.lua             # UI with fixed text wrapping
+│   │   ├── gameLogic.lua      # Game logic (fixed asset paths)
+│   │   └── gameDraw.lua       # Game rendering
+│   └── systems/               # Core systems
+│       ├── assetManager.lua   # Asset management (enhanced)
+│       └── stateManager.lua   # State management
+└── assets/                    # All assets properly organized
+    ├── images/
+    │   └── fundo.jpg          # Background image
+    ├── audio/
+    │   └── click_sound.mp3    # Click sound
+    ├── fonts/
+    │   └── PressStart2P-Regular.ttf # Game font
+    └── data/
+        ├── valid_answers.csv  # Answer words ✅
+        └── valid_guesses.csv  # Guess words ✅
+```
+
+### ✅ All Issues Resolved:
+1. ✅ **Asset paths fixed** - CSV files now found correctly
+2. ✅ **Text overflow fixed** - "Pressiona R ou ESC" wraps properly
+3. ✅ **Proper modularization** - Clean folder structure
+4. ✅ **Asset management** - Centralized asset loading
+5. ✅ **Library organization** - External libs in proper location
+
+### 🚀 Game Status: **WORKING PERFECTLY!**
+
+The game now loads without errors and all assets are properly organized following Lua/gamedev best practices while keeping it simple for a non-professional game.
+
+**Ready to play! 🎮**
